Move definition of struct LuaRef to lua_ref.hpp
